+++ src/main/java/com/edutecno/modelo/dto/CategoriaDTO.java
@@ -1,19 +1,19 @@
 package com.edutecno.modelo.dto;
 
 public class CategoriaDTO {
-	private double idCategoria;
+	private int idCategoria;
 	private String nombreCategoria;
 	
 	
-	public CategoriaDTO(double idCategoria, String nombreCategoria) {
+	public CategoriaDTO(int idCategoria, String nombreCategoria) {
 		super();
 		this.idCategoria = idCategoria;
 		this.nombreCategoria = nombreCategoria;
 	}
-	public double getIdCategoria() {
+	public int getIdCategoria() {
 		return idCategoria;
 	}
-	public void setIdCategoria(double idCategoria) {
+	public void setIdCategoria(int idCategoria) {
 		this.idCategoria = idCategoria;
 	}
 	public String getNombreCategoria() {
